requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
timm>=0.9.0

# Data & Augmentation
albumentations>=1.3.0
opencv-python-headless>=4.8.0
pillow>=10.0.0

# Metrics & Analysis
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Explainability
grad-cam>=1.4.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
